+++ web-marketplace/src/legacy-pages/Dashboard/MyCreditBatches/MyCreditBatches.Table.tsx
@@ -8,8 +8,10 @@ import PlusIcon from 'web-components/src/components/icons/PlusIcon';
 import { TablePaginationParams } from 'web-components/src/components/table/ActionsTable';
 import { Subtitle } from 'web-components/src/components/typography';
 
-import { CreateBatchOrganizationContext } from 'types/ledger/batch';
-import { BatchInfoWithSupply } from 'types/ledger/ecocredit';
+import {
+  BatchInfoWithSupply,
+  CreateBatchOrganizationContext,
+} from 'types/ledger/ecocredit';
 import { UseStateSetter } from 'types/react/use-state';
 
 import WithLoader from 'components/atoms/WithLoader';
