--- web-marketplace/src/types/ledger/batch.ts
@@ -1,11 +1,0 @@
-/**
- * Organization context for creating credit batches
- * Used when creating batches from an organization dashboard
- */
-export type CreateBatchOrganizationContext = {
-  issuerAddress?: string;
-  organizationDaoAddress?: string;
-  organizationRbamAddress?: string;
-  authorizationId?: number;
-  roleId?: number;
-};
