+++ web-marketplace/src/types/ledger/ecocredit.ts
@@ -71,3 +71,15 @@ export interface BasketTokens {
   balance?: BankQueryBalanceResponse;
   metadata?: QueryDenomMetadataResponse;
 }
+
+/**
+ * Organization context for creating credit batches
+ * Used when creating batches from an organization dashboard
+ */
+export type CreateBatchOrganizationContext = {
+  issuerAddress?: string;
+  organizationDaoAddress?: string;
+  organizationRbamAddress?: string;
+  authorizationId?: number;
+  roleId?: number;
+};
